Factor out code from ServiceSpec that can be used to control the local-cluster from integration tests
